Add support for negative indices and improve efficiency
Wavefront .obj files support vertex indices starting from the first
defined vertex, or indices relative to the most recently defined vertex.
Previously, the code would choke on negative indices, this makes it so
it doesn't.

In the process I also made a few improvements for the sake of
performance. To avoid allocating for each line, I changed the lexer to
re-use the args buffer, growing when necessary. This required some
unsafe code to handle lifetime-related issues, but the code has a number
of comments explaining what is going on.

I also changed parsing the vertex group to avoid allocation. Instead it
parses a vertex group into a triple, using '0' to indicate a missing
value.