Add support for lines and points geometry statements
They're easy enough to add support for, there was already code for them!
Most signficant change was that Lines are represented as a series of
line segments. The spec doesn't say how to interpret line statements
with more than two vertices, so this should be the most flexible
representation.

It also adds support for reading the (deprecated) "fo" statement as an
"f" statement.